refactor: break FakeVec into a module
