+++ rust/relay-core/src/lib.rs
@@ -5,13 +5,15 @@
 pub mod error;
 pub use error::*;
 
+/// FakeVec
+pub mod fake_vec;
+pub use fake_vec::*;
+
 use bitcoin_spv::{
     btcspv::retarget_algorithm,
     std_types::SPVProof,
     types::{Hash256Digest, HeaderArray, RawHeader},
 };
-
-use generic_array::{ArrayLength, GenericArray};
 
 #[repr(C)]
 #[derive(Clone, Copy, Default, Debug, PartialEq, serde::Serialize, serde::Deserialize)]
@@ -27,22 +29,6 @@ pub struct HeaderInfo {
     pub height: u32,
 }
 
-/// A simple Fake Vector. It has a fixed-sized allocation and overwrites things at the front when
-/// it runs out of space.
-///
-/// It can be used at most `usize::MAX` times. It panics after that. This shouldn't be an issue.
-#[repr(C)]
-#[derive(Debug, Clone, Default, PartialEq, Eq, serde::Serialize, serde::Deserialize)]
-#[serde(bound = "N: ArrayLength<T>")]
-pub struct FakeVec<T, N>
-where
-    T: serde::Serialize + for<'d> serde::Deserialize<'d> + Default,
-    N: ArrayLength<T>,
-{
-    latest: usize,
-    internal: GenericArray<T, N>,
-}
-
 /// A Raw header with its associated info.
 /// Convenience struct to avoid declaring twice as many let bindings
 #[repr(C)]
@@ -54,44 +40,6 @@ pub struct RawWithInfo<'a> {
     pub info: &'a HeaderInfo,
 }
 
-impl<T, N> FakeVec<T, N>
-where
-    T: serde::Serialize + for<'d> serde::Deserialize<'d> + Default,
-    N: ArrayLength<T>,
-{
-    // Get the capacity of the underlying array
-    fn capacity(&self) -> usize {
-        self.internal.len()
-    }
-
-    // Get the actual index within the array
-    fn normalize_index(&self, index: usize) -> usize {
-        index % self.capacity()
-    }
-
-    // Check if we have the item referenced by an index
-    fn valid(&self, index: usize) -> bool {
-        index <= self.latest && index > self.latest - self.capacity()
-    }
-
-    /// Push an item to the Buffer
-    pub fn push(&mut self, item: T) -> usize {
-        self.latest += 1;
-        let index = self.normalize_index(self.latest);
-        self.internal[index] = item;
-        self.latest
-    }
-
-    /// Get an item at an index, if we still have it
-    pub fn get(&self, index: usize) -> Option<&T> {
-        if !self.valid(index) {
-            None
-        } else {
-            let index = self.normalize_index(index);
-            Some(&self.internal[index])
-        }
-    }
-}
 
 #[repr(C)]
 #[derive(Clone, Debug, Default, PartialEq, serde::Serialize, serde::Deserialize)]
@@ -105,6 +53,7 @@ pub struct Relay {
     pub best_known_digest: Hash256Digest,
     /// The LCA of the most recent reorg or extension
     pub last_reorg_lca: Hash256Digest,
+    // TODO: generalize
     header_store: FakeVec<HeaderInfo, generic_array::typenum::U4096>,
 }
